Add polygonize raster after mosaicing output

+++ utils/data_processing/merge_output.py
@@ -7,8 +7,8 @@ import os
 
 
 def merge_output(shape, tiles_dir):
-    final_output = np.zeros(shape)
-    denominator = np.zeros(shape)
+    final_output = np.zeros(shape, dtype=np.uint8)
+    denominator = np.zeros(shape, dtype=np.uint8)
     for ele in os.listdir(tiles_dir):
         left, down, right, top = ele.split("_")[-4:]
         top = top.split(".")[0]
@@ -17,4 +17,4 @@ def merge_output(shape, tiles_dir):
         denominator[int(left) : int(right), int(down) : int(top)] += 1
     denominator[denominator == 0] = 1
     final_output = final_output / denominator
-    return final_output
+    return final_output.astype(np.uint8)
